git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@18141 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/bus_managers/acpi/resources/rscalc.c
@@ -1,7 +1,7 @@
 /*******************************************************************************
  *
  * Module Name: rscalc - Calculate stream and list lengths
- *              $Revision: 1.78 $
+ *              $Revision: 1.79 $
  *
  ******************************************************************************/
 
@@ -557,7 +557,7 @@ AcpiRsGetListLength (
          */
         BufferSize = AcpiGbl_ResourceStructSizes[ResourceIndex] +
                         ExtraStructBytes;
-        BufferSize = ACPI_ROUND_UP_TO_NATIVE_WORD (BufferSize);
+        BufferSize = (UINT32) ACPI_ROUND_UP_TO_NATIVE_WORD (BufferSize);
 
         *SizeNeeded += BufferSize;
